NoteController.java

+++ backend/src/main/java/com/campusnote/controller/NoteController.java
@@ -1,17 +1,24 @@
 package com.campusnote.controller;
 
+import java.util.List;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
 import com.campusnote.dto.NoteRequest;
 import com.campusnote.dto.NoteResponse;
 import com.campusnote.service.NoteService;
 
 import jakarta.validation.Valid;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/notes")
@@ -45,7 +52,7 @@ public class NoteController {
 
     @GetMapping("/{id}")
     public ResponseEntity<NoteResponse> getNoteById(
-            @PathVariable Long id) {
+            @PathVariable Integer id) {
 
         return ResponseEntity.ok(
                 noteService.getNoteById(id)
@@ -55,7 +62,7 @@ public class NoteController {
     @GetMapping("/subject/{subjectId}")
     public ResponseEntity<List<NoteResponse>>
     getNotesBySubject(
-            @PathVariable Long subjectId) {
+            @PathVariable Integer subjectId) {
 
         return ResponseEntity.ok(
                 noteService.getNotesBySubject(
@@ -67,7 +74,7 @@ public class NoteController {
     @GetMapping("/user/{userId}")
     public ResponseEntity<List<NoteResponse>>
     getNotesByUser(
-            @PathVariable Long userId) {
+            @PathVariable Integer userId) {
 
         return ResponseEntity.ok(
                 noteService.getNotesByUser(
@@ -88,7 +95,7 @@ public class NoteController {
 
     @PutMapping("/{id}")
     public ResponseEntity<NoteResponse> updateNote(
-            @PathVariable Long id,
+            @PathVariable Integer id,
             @Valid @RequestBody NoteRequest request) {
 
         return ResponseEntity.ok(
@@ -98,7 +105,7 @@ public class NoteController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteNote(
-            @PathVariable Long id) {
+            @PathVariable Integer id) {
 
         noteService.deleteNote(id);
